Modify generatehistogram.py, realdata.py and README.md. Add code for calibration concordance.ipynb
